+++ lib/docx_template.dart
@@ -1,5 +1,14 @@
 library docx_template;
 
-export 'src/template.dart' show DocxTemplate, TagPolicy, ImagePolicy;
+export 'src/template.dart'
+    show DocxTemplate, DocxTemplateException, TagPolicy, ImagePolicy;
 export 'src/model.dart';
 export 'src/tag_models.dart' show TagType, DocxTag, DocxTagCollection;
+export 'src/editor.dart'
+    show
+        DocxEditableStructure,
+        DocxParagraphInfo,
+        DocxTableInfo,
+        TemplatedCell,
+        TemplatedRow,
+        SdtIdAllocator;